+++ demisto_sdk/commands/validate/validate_manager.py
@@ -102,8 +102,8 @@ from demisto_sdk.commands.common.tools import (
     _get_file_id, find_type, get_api_module_ids,
     get_api_module_integrations_set, get_content_path, get_file,
     get_pack_ignore_file_path, get_pack_name, get_pack_names_from_files,
-    get_relative_path_from_packs_dir, get_yaml, open_id_set_file,
-    run_command_os)
+    get_relative_path_from_packs_dir, get_remote_file, get_yaml,
+    open_id_set_file, run_command_os)
 from demisto_sdk.commands.create_id_set.create_id_set import IDSetCreator
 
 SKIPPED_FILES = ['CommonServerPython.py', 'CommonServerUserPython.py', 'demistomock.py']
@@ -859,9 +859,9 @@ class ValidateManager:
 
         validation_results.add(self.validate_modified_files(modified_files))
         validation_results.add(self.validate_added_files(added_files, modified_files))
+        validation_results.add(self.validate_deleted_files(deleted_files, added_files))
         validation_results.add(self.validate_changed_packs_unique_files(modified_files, added_files, old_format_files,
                                                                         changed_meta_files))
-        validation_results.add(self.validate_deleted_files(deleted_files, added_files))
 
         if old_format_files:
             click.secho(f'\n================= Running validation on old format files =================',
@@ -1345,11 +1345,12 @@ class ValidateManager:
 
         """
         file_path = str(file_path)
-        file_type = find_type(file_path)
+        file_dict = get_remote_file(file_path)
+        file_type = find_type(file_path, file_dict)
         return file_type in FileType_ALLOWED_TO_DELETE or not file_type
 
     @staticmethod
-    def was_file_renamed_but_labeled_as_deleted(file_path, added_files):
+    def was_file_renamed_but_labeled_as_deleted(deleted_file_path, added_files):
         """ Check if a file was renamed and not deleted (git false label the file as deleted)
         Args:
             file_path: The file path.
@@ -1358,15 +1359,16 @@ class ValidateManager:
 
         """
         if added_files:
-            file_path = str(file_path)
-            if file_type := find_type(file_path):
-                deleted_file_dict = get_file(file_path, file_type)
-                deleted_file_id = _get_file_id(file_path, deleted_file_dict)
+            deleted_file_path = str(deleted_file_path)
+            deleted_file_dict = get_remote_file(deleted_file_path)
+            if deleted_file_type := find_type(deleted_file_path, deleted_file_dict):
+                deleted_file_id = _get_file_id(deleted_file_type.value, deleted_file_dict)
                 if deleted_file_id:
                     for file in added_files:
                         file = str(file)
-                        file_dict = get_file(file, find_type(file))
-                        if deleted_file_id == _get_file_id(file, file_dict):
+                        file_type = find_type(file)
+                        file_dict = get_file(file, file_type.value)
+                        if deleted_file_id == _get_file_id(file_type.value, file_dict):
                             return True
         return False
 
@@ -1390,7 +1392,7 @@ class ValidateManager:
                     if not self.is_file_allowed_to_be_deleted(file_path):
                         file_path = str(file_path)
                         error_message, error_code = Errors.file_cannot_be_deleted(file_path)
-                        if self.handle_error(error_message, error_code, file_path):
+                        if self.handle_error(error_message, error_code, ''):
                             is_valid = False
 
         return is_valid