+++ CHANGELOG.md
@@ -61,6 +61,7 @@
 * Locked the dependency on Docker.
 * Removed a traceback line from the **init** command templates: BaseIntegration, BaseScript.
 * Updated the token in **_add_pr_comment** method from the content-bot token to the xsoar-bot token.
+* Fixed an issue in the **validate** command when validating deleted files.
 
 ## 1.6.7